refactor: fix typing errors (pyright) (#855)

+++ memgpt/cli/cli.py
@@ -28,7 +28,7 @@ from memgpt.constants import MEMGPT_DIR, CLI_WARNING_PREFIX, JSON_ENSURE_ASCII
 from memgpt.agent import Agent
 from memgpt.embeddings import embedding_model
 from memgpt.server.constants import WS_DEFAULT_PORT, REST_DEFAULT_PORT
-from memgpt.data_types import AgentState, LLMConfig, EmbeddingConfig, User
+from memgpt.data_types import AgentState, LLMConfig, EmbeddingConfig, User, Passage
 from memgpt.metadata import MetadataStore, save_agent
 from memgpt.migrate import migrate_all_agents, migrate_all_sources
 
@@ -325,19 +325,21 @@ def server(
         # Run the command
         typer.secho(f"Running WS (websockets) server: {command} (inside {server_directory})")
 
+        process = None
         try:
             # Start the subprocess in a new session
             process = subprocess.Popen(command, shell=True, start_new_session=True, cwd=server_directory)
             process.wait()
         except KeyboardInterrupt:
             # Handle CTRL-C
-            typer.secho("Terminating the server...")
-            process.terminate()
-            try:
-                process.wait(timeout=5)
-            except subprocess.TimeoutExpired:
-                process.kill()
-                typer.secho("Server terminated with kill()")
+            if process is not None:
+                typer.secho("Terminating the server...")
+                process.terminate()
+                try:
+                    process.wait(timeout=5)
+                except subprocess.TimeoutExpired:
+                    process.kill()
+                    typer.secho("Server terminated with kill()")
             sys.exit(0)
 
 
@@ -463,11 +465,6 @@ def run(
     ms = MetadataStore(config)
     user = create_default_user_or_exit(config, ms)
 
-    # override with command line arguments
-    if debug:
-        config.debug = debug
-    if no_verify:
-        config.no_verify = no_verify
     # determine agent to use, if not provided
     if not yes and not agent:
         agents = ms.list_agents(user_id=user.id)
@@ -482,10 +479,11 @@ def run(
                 agent = questionary.select("Select agent:", choices=agents).ask()
 
     # create agent config
-    if agent and ms.get_agent(agent_name=agent, user_id=user.id):  # use existing agent
+    agent_state = ms.get_agent(agent_name=agent, user_id=user.id) if agent else None
+    if agent and agent_state:  # use existing agent
         typer.secho(f"\n🔁 Using existing agent {agent}", fg=typer.colors.GREEN)
         # agent_config = AgentConfig.load(agent)
-        agent_state = ms.get_agent(agent_name=agent, user_id=user.id)
+        # agent_state = ms.get_agent(agent_name=agent, user_id=user_id)
         printd("Loading agent state:", agent_state.id)
         printd("Agent state:", agent_state.state)
         # printd("State path:", agent_config.save_state_dir())
@@ -537,7 +535,7 @@ def run(
         ms.update_agent(agent_state)
 
         # create agent
-        memgpt_agent = Agent(agent_state, interface=interface)
+        memgpt_agent = Agent(agent_state, interface=interface())
 
     else:  # create new agent
         # create new agent config: override defaults with args if provided
@@ -605,7 +603,7 @@ def run(
         try:
             memgpt_agent = presets.create_agent_from_preset(
                 agent_state=agent_state,
-                interface=interface,
+                interface=interface(),
             )
             save_agent(agent=memgpt_agent, ms=ms)
         except ValueError as e:
@@ -690,7 +688,7 @@ def delete_agent(
 
 def attach(
     agent_name: Annotated[str, typer.Option(help="Specify agent to attach data to")],
-    data_source: Annotated[str, typer.Option(help="Data source to attach to avent")],
+    data_source: Annotated[str, typer.Option(help="Data source to attach to agent")],
     user_id: uuid.UUID = None,
 ):
     # use client ID is no user_id provided
@@ -704,7 +702,7 @@ def attach(
 
         ms = MetadataStore(config)
         agent = ms.get_agent(agent_name=agent_name, user_id=user_id)
-        print(agent.id, agent.user_id, user_id)
+        assert agent is not None, f"No agent found under agent_name={agent_name}, user_id={user_id}"
         source = ms.get_source(source_name=data_source, user_id=user_id)
         assert source is not None, f"Source {data_source} does not exist for user {user_id}"
 
@@ -723,6 +721,7 @@ def attach(
 
             # need to associated passage with agent (for filtering)
             for passage in passages:
+                assert isinstance(passage, Passage), f"Generate yielded bad non-Passage type: {type(passage)}"
                 passage.agent_id = agent.id
 
             # insert into agent archival memory
@@ -735,7 +734,9 @@ def attach(
         dest_storage.save()
 
         # attach to agent
-        source_id = ms.get_source(source_name=data_source, user_id=user_id).id
+        source = ms.get_source(source_name=data_source, user_id=user_id)
+        assert source is not None, f"source does not exist for source_name={data_source}, user_id={user_id}"
+        source_id = source.id
         ms.attach_source(agent_id=agent.id, source_id=source_id, user_id=user_id)
 
         total_agent_passages = dest_storage.size()
